Remove the GHC.SourceGen.Syntax module.
Instead, export types from corresponding modules (e.g.: `HsExpr'`
from `GHC.SourceGen.Expr`.  Haddock does a decent job of wiring
everything up the right way.
